+++ kolena/_experimental/object_detection/__init__.py
@@ -29,7 +29,7 @@ from .workflow import ThresholdConfiguration
 
 from .evaluator import ObjectDetectionEvaluator
 from .dataset import upload_object_detection_results
-from .dataset import labeled_bounding_box_as_dict
+from .utils import labeled_bounding_box_as_dict
 
 __all__ = [
     "TestSample",
